Update ECB api base url
SDW web services API has been moved to the European Data Portal and the url now redirects to `https://data-api.ecb.europa.eu`. This changes the base url to the new one. This hopefully also fixes CORS errors as it seems that the CORS headers were dropped in the redirect process.